FEAT : 썸네일 및 썸머노트 이미지 업로드 구현
보완해야 할 점
1. 유효성 체크 수정 필요
2. SummerNote 이미지 업로드 문제점
   2.1. SummnerNote는 이미지 업로드 시 ajax 비동기 방식을 사용
   2.2. 업로드 후 업로드된 자원 url을 GET요청하여 img태그 생성해줌
   2.3. 그러나 2.1 과정에서 정상적으로 업로드 된 이미지가 인텔리제이에 동기화가 바로 안 됨
   2.4. 따라서 2.2 과정 수행 시 404 Not Found 오류가 발생됨
   2.5. 이미지 자원을 GET방식으로 다시 요청, setTimeout을 해봤지만 해결되지 않음

+++ src/main/java/com/kh/learnovation/domain/course/dto/CourseImageDTO.java
@@ -18,6 +18,6 @@ public class CourseImageDTO {
     private String originalImageName;
     private String savedImageName;
     private String savedPath;
-    private int imageSize;
+    private Long imageSize;
     private Timestamp createAt;
 }
